requirements-cs5914.txt
streamlit==1.26.0
openai==0.28.0
toml~=0.10.2
pandas==2.1.0
tabulate==0.9.0
python-dotenv==1.0.0
streamlit-ace
patch~=1.16
unidiff~=0.7.5
pyperclip~=1.8.2
httpx==0.25.0
st_ant_tree==0.0.4
pyautogen==0.1.13
requests~=2.31.0
asyncio
pytest
boto3

